requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# cardiogram_experiment/create_grayscale_images.py: 7
# shapes_experiment/stimuli/draw.py: 3
# shapes_v2_experiment/create_stimuli.py: 11
Pillow == 3.3.1

# cardiogram_experiment/exemplar_model.py: 11
# cardiogram_experiment/graph_accuracies.py: 6
# cardiogram_experiment/graph_layer_0.py: 13
# cog_dev_experiment/accuracy.py: 8
# examining_weights/figures.py: 5
# gabor_experiment/figure.py: 9
# gabor_experiment/pca.py: 22
# gabor_experiment/sim.py: 15
# shapes_experiment/luce_figures.py: 5
# shapes_experiment/misc.py: 8
# shapes_experiment/optimum_figures.py: 7
# shapes_experiment/other_figures.py: 5
# shapes_v2_experiment/better_figures.py: 5
# shapes_v2_experiment/figures.py: 7
matplotlib == 1.5.3

# cardiogram_experiment/exemplar_model.py: 8
# cardiogram_experiment/graph_accuracies.py: 4
# cardiogram_experiment/graph_layer_0.py: 10
# cardiogram_experiment/misc.py: 5
# cardiogram_experiment/net.py: 44
# cog_dev_experiment/accuracy.py: 5
# cog_dev_experiment/net.py: 45
# examining_weights/analyse_weights.py: 6
# examining_weights/exemplar_model_circle_square_bounding_box.py: 6
# examining_weights/get_weights.py: 3
# gabor_experiment/figure.py: 5
# gabor_experiment/pca.py: 18
# gabor_experiment/sim.py: 11
# shapes_experiment/exemplar_model_circle_square.py: 6
# shapes_experiment/exemplar_model_circle_square_bounding_box.py: 6
# shapes_experiment/luce_figures.py: 3
# shapes_experiment/misc.py: 5
# shapes_experiment/net.py: 44
# shapes_experiment/optimum_figures.py: 4
# shapes_v2_experiment/create_stimuli.py: 8
# shapes_v2_experiment/exemplar_model.py: 9
# shapes_v2_experiment/figures.py: 4
# shapes_v2_experiment/net.py: 46
# tests/test_cardiogram_experiment.py: 5
# utils/iv3.py: 48
# utils/misc.py: 1
numpy == 1.11.3

# cardiogram_experiment/exemplar_model.py: 9
# cardiogram_experiment/graph_layer_0.py: 12
# cardiogram_experiment/net.py: 46
# cog_dev_experiment/accuracy.py: 6
# cog_dev_experiment/net.py: 47
# examining_weights/analyse_weights.py: 7
# examining_weights/exemplar_model_circle_square_bounding_box.py: 7
# examining_weights/figures.py: 3
# examining_weights/get_weights.py: 5
# gabor_experiment/figure.py: 6
# gabor_experiment/net.py: 44
# gabor_experiment/pca.py: 19
# gabor_experiment/sim.py: 12
# shapes_experiment/exemplar_model_circle_square.py: 7
# shapes_experiment/exemplar_model_circle_square_bounding_box.py: 7
# shapes_experiment/misc.py: 6
# shapes_experiment/net.py: 46
# shapes_experiment/optimum_figures.py: 5
# shapes_experiment/other_figures.py: 3
# shapes_v2_experiment/better_figures.py: 3
# shapes_v2_experiment/create_stimuli.py: 9
# shapes_v2_experiment/exemplar_model.py: 10
# shapes_v2_experiment/figures.py: 5
# shapes_v2_experiment/net.py: 48
# tests/test_cardiogram_experiment.py: 6
pandas == 0.18.1

# gabor_experiment/pca.py: 23
scikit_learn == 0.18.1

# cardiogram_experiment/exemplar_model.py: 10
# cardiogram_experiment/graph_accuracies.py: 5
# cardiogram_experiment/graph_layer_0.py: 11
# cog_dev_experiment/accuracy.py: 7
# examining_weights/exemplar_model_circle_square_bounding_box.py: 8
# examining_weights/figures.py: 4
# gabor_experiment/figure.py: 7
# gabor_experiment/pca.py: 20
# gabor_experiment/sim.py: 13
# shapes_experiment/exemplar_model_circle_square.py: 8
# shapes_experiment/exemplar_model_circle_square_bounding_box.py: 8
# shapes_experiment/luce_figures.py: 4
# shapes_experiment/misc.py: 7
# shapes_experiment/optimum_figures.py: 6
# shapes_experiment/other_figures.py: 4
# shapes_v2_experiment/better_figures.py: 4
# shapes_v2_experiment/exemplar_model.py: 11
# shapes_v2_experiment/figures.py: 6
seaborn == 0.8.1

# utils/iv3.py: 51
six == 1.10.0

# cardiogram_experiment/net.py: 45
# cog_dev_experiment/net.py: 46
# examining_weights/get_weights.py: 4
# gabor_experiment/net.py: 45
# shapes_experiment/net.py: 45
# shapes_v2_experiment/net.py: 47
# utils/iv3.py: 49
tensorflow >= 0.11.0
